+++ tests/test_cache.py
@@ -2,32 +2,19 @@ from pathlib import Path
 
 from kaizo import ConfigParser
 
-cache_py = """
-num_call = 0
-def fn():
-    global num_call
-
-    num_call+=1
-    return num_call
-"""
-
 cache_config = """
-local: main.py
 with_cache:
-  module: local
-  source: fn
+  module: random
+  source: random
   cache: true
 without_cache:
-  module: local
-  source: fn
+  module: random
+  source: random
   cache: false
 """
 
 
 def test_cache(tmp_path: Path) -> None:
-    module = tmp_path / "main.py"
-    module.write_text(cache_py)
-
     cfg_file = tmp_path / "cfg.yml"
     cfg_file.write_text(cache_config)
 
@@ -40,4 +27,4 @@ def test_cache(tmp_path: Path) -> None:
     without_cache_2 = out["without_cache"]
 
     assert with_cache_1 == with_cache_2
-    assert without_cache_2 == without_cache_1 + 1
+    assert without_cache_2 != without_cache_1